backend/requirements.txt
# Requirements for backend
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# ML/NLP
sentence-transformers>=3.0.0
scikit-learn==1.3.2
numpy==1.24.3

# LLM Integration
# openai==1.3.0  # Uncomment for OpenAI
requests==2.31.0

# Database (optional)
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9
# alembic==1.13.0

# Vector DB (optional)
# pinecone-client==3.0.0
# redis==5.0.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.12.0
flake8==6.1.0
mypy==1.7.1
